+++ backend/canisters/user/api/src/main.rs
@@ -7,13 +7,14 @@ fn main() {
     generate_candid_method!(user, events, query);
     generate_candid_method!(user, events_by_index, query);
     generate_candid_method!(user, events_window, query);
+    generate_candid_method!(user, hot_group_exclusions, query);
     generate_candid_method!(user, initial_state_v2, query);
     generate_candid_method!(user, messages_by_message_index, query);
     generate_candid_method!(user, public_profile, query);
     generate_candid_method!(user, search_messages, query);
     generate_candid_method!(user, updates_v2, query);
 
-    generate_candid_method!(user, add_recommended_group_exclusions, update);
+    generate_candid_method!(user, add_hot_group_exclusions, update);
     generate_candid_method!(user, add_reaction, update);
     generate_candid_method!(user, archive_chat, update);
     generate_candid_method!(user, assume_group_super_admin, update);