+++ backend/canisters/user/impl/src/model/mod.rs
@@ -4,5 +4,5 @@ pub mod direct_chat;
 pub mod direct_chats;
 pub mod group_chat;
 pub mod group_chats;
-pub mod recommended_group_exclusions;
+pub mod hot_group_exclusions;
 pub mod unread_message_index_map;