+++ backend/canisters/user_index/impl/src/guards.rs
@@ -24,8 +24,8 @@ pub fn caller_is_group_index() -> Result<(), String> {
     }
 }
 
-pub fn caller_is_super_admin() -> Result<(), String> {
-    if read_state(|state| state.is_caller_super_admin()) {
+pub fn caller_is_platform_moderator() -> Result<(), String> {
+    if read_state(|state| state.is_caller_platform_moderator()) {
         Ok(())
     } else {
         Err("Caller is not a super admin".to_string())