+++ backend/canisters/user_index/impl/src/jobs/mod.rs
@@ -1,6 +1,6 @@
 use crate::RuntimeState;
 
-pub mod dismiss_super_admins;
+pub mod dismiss_platform_moderators;
 pub mod notify_user_principal_migrations;
 pub mod set_users_suspended;
 pub mod sync_events_to_local_user_index_canisters;
@@ -8,7 +8,7 @@ pub mod sync_users_to_storage_index;
 pub mod upgrade_canisters;
 
 pub(crate) fn start(runtime_state: &RuntimeState) {
-    dismiss_super_admins::start_job_if_required(runtime_state);
+    dismiss_platform_moderators::start_job_if_required(runtime_state);
     notify_user_principal_migrations::start_job_if_required(runtime_state);
     set_users_suspended::start_job_if_required(runtime_state);
     sync_events_to_local_user_index_canisters::start_job_if_required(runtime_state);