+++ backend/canisters/user_index/impl/src/updates/remove_platform_moderator.rs
@@ -5,13 +5,13 @@ use canister_tracing_macros::trace;
 use local_user_index_canister::{Event, SuperAdminStatusChanged};
 use types::{ChatId, UserId};
 use user_canister::c2c_revoke_super_admin;
-use user_index_canister::remove_super_admin::{Response::*, *};
+use user_index_canister::remove_platform_moderator::{Response::*, *};
 
 #[proposal(guard = "caller_is_governance_principal")]
 #[trace]
-async fn remove_super_admin(args: Args) -> Response {
-    if !read_state(|state| is_already_super_admin(&args.user_id, state)) {
-        return NotSuperAdmin;
+async fn remove_platform_moderator(args: Args) -> Response {
+    if !read_state(|state| is_already_platform_moderator(&args.user_id, state)) {
+        return NotPlatformModerator;
     }
 
     let c2c_args = c2c_revoke_super_admin::Args {};
@@ -25,17 +25,20 @@ async fn remove_super_admin(args: Args) -> Response {
     }
 }
 
-fn is_already_super_admin(user_id: &UserId, runtime_state: &RuntimeState) -> bool {
-    runtime_state.data.super_admins.contains(user_id)
+fn is_already_platform_moderator(user_id: &UserId, runtime_state: &RuntimeState) -> bool {
+    runtime_state.data.platform_moderators.contains(user_id)
 }
 
 fn commit(user_id: &UserId, groups_to_dismiss_user_from: Vec<ChatId>, runtime_state: &mut RuntimeState) {
-    runtime_state.data.super_admins.remove(user_id);
+    runtime_state.data.platform_moderators.remove(user_id);
 
     for group_id in groups_to_dismiss_user_from {
-        runtime_state.data.super_admins_to_dismiss.push_back((*user_id, group_id));
+        runtime_state
+            .data
+            .platform_moderators_to_dismiss
+            .push_back((*user_id, group_id));
     }
-    crate::jobs::dismiss_super_admins::start_job_if_required(runtime_state);
+    crate::jobs::dismiss_platform_moderators::start_job_if_required(runtime_state);
 
     runtime_state.data.push_event_to_all_local_user_indexes(
         Event::SuperAdminStatusChanged(SuperAdminStatusChanged {