+++ backend/integration_tests/src/client/group_index.rs
@@ -8,5 +8,4 @@ generate_query_call!(search);
 generate_update_call!(add_local_group_index_canister);
 generate_update_call!(delete_frozen_group);
 generate_update_call!(freeze_group);
-generate_update_call!(reinstall_group);
 generate_update_call!(unfreeze_group);