+++ frontend/app/src/components/home/proposals/ProposalContent.svelte
@@ -408,5 +408,6 @@
 
     .payload-body {
         word-wrap: break-word;
+        max-height: 65vh;
     }
 </style>